Major changes:
Removed Material UI and refactored app with Bootstrap elements. This changes the UI drastically, making it easier for users to log their workouts, create posts, as well as creating a much better user experience.

Reconfigured create post, create workout, account and home page (community) with a new design. While some Material UI may exist, it will be be completely removed in a future update.

New Features:

When a user chooses to log a run or a walk, it will give them an option for miles and pace instead of weight and rep. Similarly, when a user logs a work out with a weight type activity it will give them weights as well as reps. In other words, the workouts now correspond with the type of data.

Removed workout history text when user adds a workout. This was present so the user can see their last set as they add new ones, but was not user friendly. This was to keep the UI cleaner and not over cluttered. Instead, now when a user adds a set, it will log the previous set into local storage so even if the user logs out, refreshes the page, losses service, or leaves the page, their activity, set, reps will still be present and will clear once their workout is logged.

Added ‘add set’ and ‘remove set’ instead of sets populating once added. This will save users data locally each time.

Added “remove set” so now a user can remove a set if they so chose too, as it will delete from newest added sets, to oldest. The user cannot delete all sets, as a warning will be prompt indicating the user must have at least once active set. Use the user chooses to remove a set, it will also remove from local storage

Now when a user logs a workout, it will appear on the log workout page so they can look back and see their history

Added new text to appear if user is not logged in or does not have an account and wishes to leave a comment.

Temporarily removed “notes” from each logged workout as this will be added in another location to keep UI clean.

Fixes:

Fixed issue where if a user was not logged and clicking on account would throw an error. Made it so the account is not available if a user is not logged in, but furthermore created an additional step to redirect anyone who's not logged into the login page if they happen to go to the account URL.

Fixed issue where adding a workout the set would auto increment instead of the users input.

Other:

Removed a lot of notes and unnecessary code and used prettier to clean up code. Added a lot of new css elements and features to the app. Temporarily removed notes from the syntax to make the code easier to read when coding. Will add them back once app is more or less fully functioning.

Future:

Create a cleaner UI for user when editing their workouts /or viewing their workouts

Add overall weight lifted, miles walked/ran, overall page, as well as time spent for each workout.

Like comments.

Reply to comments.

Add reactions to posts.

Favorite Posts