Main features added:
Added React Quill to components that used text. Now users can edit their text when posting and when editing on their profile or home page.
Added search bar to navbar. This search will only look for keywords in the content of the body and the title. The search is works where ever their are posts as it was added to main.jsx. A future update will allow users to filter down posts.
Added user options to every post. Now a user can choose to favorite, react, view comments or share a post. As of now, these features are still in the process
Added the ability for uses to toggle comments by viewing them. This way the page is kept cleaner. A future update will allow the user to open the post on a separate page.

Other changes:
Made various changes to some syntax and removed unnecessary code that was no longer needed.
Adjusted the placements of the edit and delete icons
Removed main toastify from the jsx and now only have it in App.jsx. Temp removed Toast css for troubleshooting but will add back later.