New Features Added:
Users can now favorite a post. A user can toggle between keeping a post as their favroites and all favorites will appear on their profile. As of now, favroites do not redirect to the post route but will in an upcoming feature

Users can now reply to comments. A user can now reply to any comment as well as have the ability to toggle between viewing or not viewing all the replies. The replies will indicting who the users is replying too.

Added error boundary: If a user comes across an error, they will get a page error instead of one from react. This is a boilerplate code.

Added icons.

Other:

Fixed some CSS elements that were not working correctly.
Removed a lot of code that was either redundant or not needed.
Adjusted typedefs to meet standards.
Adjusted queries to meet standards
Made other small adjustments.